Added script to connect to database and imported where appropriate
